refactor : ColorKey 타입 위치를 constants/color로 변경
