--- frontend/src/types/colorKey.ts
@@ -1,3 +1,0 @@
-import { COLOR_MAP } from '@/constants/colorMap';
-
-export type ColorKey = keyof typeof COLOR_MAP;
